Move export dashboards cmd into own file
